Create README.md
In this stage of work, we carried out an exploratory analysis of Brazilian demographic data. The data used for the analysis were made available within the “demography” branch. For the code in the .rmd file to work, these files must be associated with the working directory. The two .rmd files refer to the same text, one in English, the other in Portuguese (Brazil).