Create README.md
The Doctoral Thesis was divided into work stages. Each of the steps represents a branch associated with the research performed.

The "Land Use and Cover" and "Demography" stages represent the exploratory analyzes of the data provided by the Brazilian Institute of Geography and Statistics (IBGE) and can be accessed through the institute's digital portal (https://www.ibge.gov. br/). The branches contain all the files necessary to run the code produced in R MarkDown. Note that all files associated with both exploratory analyzes must be in the working directory.